Move test utils around
