+++ builtin/v9/migration/test/util.go
@@ -2,10 +2,13 @@ package migration
 
 import (
 	"context"
+	"crypto/sha256"
+	"errors"
 	"fmt"
 	"sync"
-
 	"testing"
+
+	"github.com/filecoin-project/go-state-types/rt"
 
 	block "github.com/ipfs/go-block-format"
 	ipldcbor "github.com/ipfs/go-ipld-cbor"
@@ -199,4 +202,40 @@ func (ss *SyncBlockStoreInMemory) Put(ctx context.Context, b block.Block) error 
 	ss.mu.Lock()
 	defer ss.mu.Unlock()
 	return ss.bs.Put(ctx, b)
+}
+
+type TestLogger struct {
+	TB testing.TB
+}
+
+func (t TestLogger) Log(_ rt.LogLevel, msg string, args ...interface{}) {
+	t.TB.Logf(msg, args...)
+}
+
+func MakeCID(input string, prefix *cid.Prefix) cid.Cid {
+	data := []byte(input)
+	if prefix == nil {
+		c, err := abi.CidBuilder.Sum(data)
+		if err != nil {
+			panic(err)
+		}
+		return c
+	}
+	c, err := prefix.Sum(data)
+	switch {
+	case errors.Is(err, mh.ErrSumNotSupported):
+		// multihash library doesn't support this hash function.
+		// just fake it.
+	case err == nil:
+		return c
+	default:
+		panic(err)
+	}
+
+	sum := sha256.Sum256(data)
+	hash, err := mh.Encode(sum[:], prefix.MhType)
+	if err != nil {
+		panic(err)
+	}
+	return cid.NewCidV1(prefix.Codec, hash)
 }
