+++ builtin/v9/migration/util.go
@@ -1,15 +1,8 @@
 package migration
 
 import (
-	"crypto/sha256"
-	"errors"
 	"sync"
-	"testing"
 
-	"github.com/filecoin-project/go-state-types/abi"
-	mh "github.com/multiformats/go-multihash"
-
-	"github.com/filecoin-project/go-state-types/rt"
 	"github.com/ipfs/go-cid"
 	"golang.org/x/xerrors"
 )
@@ -67,40 +60,4 @@ func (m *MemMigrationCache) Update(other *MemMigrationCache) {
 		m.MigrationMap.Store(key, value)
 		return true
 	})
-}
-
-type TestLogger struct {
-	TB testing.TB
-}
-
-func (t TestLogger) Log(_ rt.LogLevel, msg string, args ...interface{}) {
-	t.TB.Logf(msg, args...)
-}
-
-func MakeCID(input string, prefix *cid.Prefix) cid.Cid {
-	data := []byte(input)
-	if prefix == nil {
-		c, err := abi.CidBuilder.Sum(data)
-		if err != nil {
-			panic(err)
-		}
-		return c
-	}
-	c, err := prefix.Sum(data)
-	switch {
-	case errors.Is(err, mh.ErrSumNotSupported):
-		// multihash library doesn't support this hash function.
-		// just fake it.
-	case err == nil:
-		return c
-	default:
-		panic(err)
-	}
-
-	sum := sha256.Sum256(data)
-	hash, err := mh.Encode(sum[:], prefix.MhType)
-	if err != nil {
-		panic(err)
-	}
-	return cid.NewCidV1(prefix.Codec, hash)
 }
